Allow user to specify L2 norms
There may be instances where the weight given to the L2 norm should
be different than the spectral norm (for instance if the step size
appears to be too large and prevents convergence).

This commit allows the user to specify the L2 norm for some
(but not necessarily all) linear constraints.